Improve QKV input_injection computation
Profiler revealed that conv1d qkv computation of input injection was
actually fairly inefficient. The builtin huggingface attention and QKV
computation is more efficient. We improve this by shifting QKV
computation in DeqBertSelfAttention to its own function.
This change will allow the replacement of conv1d used to precompute
the QKV matrix for the input injection. Now, we can use the existing qkv
layers in DeqBertSelfAttention for computation of the input injection.
This offers a speedup of ~1300ms second per step.