Add caching of individual pile splits
Currently, the problem is that if the internet cuts out for a while
during download, it cancels the whole 815GB dataset download, and you
have to start from scratch. Downloading each as individual datasets
allows the caching to be complete and registered by huggingface
datasets, so when it does crash, huggingface will load the cached splits
instead.